+++ Movie/Movie/Router/RouterMain.swift
@@ -1,0 +1,10 @@
+// RouterMain.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Базовый роутер
+protocol RouterMain {
+    var navigationController: UINavigationController? { get set }
+    var assemblyBuilder: AssemblyBilderProtocol? { get set }
+}
